kvm: refactor kvm and kvm_gen
- move autogenerated code from kvm_gen to kvm
- sanity.rs is now unit test inside a module of the kvm crate
